Nib coding changes.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/branches/NibCompatibility@22874 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSScrollView.m
@@ -1270,7 +1270,7 @@ static float scrollerWidth;
 
       if ([aDecoder containsValueForKey: @"NSsFlags"])
         {
-	  unsigned int flags = [aDecoder decodeIntForKey: @"NSsFlags"];
+	  unsigned long flags = [aDecoder decodeIntForKey: @"NSsFlags"];
 	  GSScrollViewFlags scrollViewFlags;
 	  memcpy((void *)&scrollViewFlags,(void *)&flags,sizeof(struct _scrollViewFlags));
